requirements.txt
# ─── Core ────────────────────────────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# ─── HuggingFace ─────────────────────────────────────────────────────────────
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0
peft>=0.9.0
safetensors>=0.4.2
huggingface-hub>=0.21.0
datasets>=2.17.0

# ─── Image processing ────────────────────────────────────────────────────────
Pillow>=10.2.0
opencv-python>=4.9.0
albumentations>=1.3.1
scikit-image>=0.22.0

# ─── Face detection ──────────────────────────────────────────────────────────
facenet-pytorch>=2.5.3      # MTCNN для детекции лиц

# ─── Metrics ─────────────────────────────────────────────────────────────────
pytorch-fid>=0.3.0
lpips>=0.1.4
torchmetrics[image]>=1.3.0

# ─── Logging & UI ────────────────────────────────────────────────────────────
wandb>=0.16.0
tensorboard>=2.15.0
tqdm>=4.66.0
rich>=13.7.0

# ─── Utils ───────────────────────────────────────────────────────────────────
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
scipy>=1.12.0
omegaconf>=2.3.0
gdown>=5.1.0

# ─── Optional: Gradio UI ─────────────────────────────────────────────────────
gradio>=4.19.0

# ─── Optional: 8-bit optimizer (экономит VRAM) ───────────────────────────────
bitsandbytes>=0.42.0
